Refactor controller and repository
