+++ src/main/java/com/lecturesearch/lecture/OAuth2/repository/UserRepository.java
@@ -1,5 +1,6 @@
-package com.lecturesearch.lecture.OAuth2;
+package com.lecturesearch.lecture.OAuth2.repository;
 
+import com.lecturesearch.lecture.OAuth2.User;
 import org.springframework.data.jpa.repository.JpaRepository;
 
 public interface UserRepository extends JpaRepository<User, Long> {
