+++ src/main/resources/application.yml
@@ -8,5 +8,5 @@ spring:
               client-secret:
           google:
             client-id:
-            client-secret: 
+            client-secret:
