+++ src/main/resources/templates/form.html
@@ -70,7 +70,7 @@
             <table style="text-align: center; align-content: center; margin: auto;">
                 <tr>
                     <td style="width: 110px;"><label for="signupEmail">Email</label></td>
-                    <td style="width: 250px;"><input type="text" class="form-control" id="signupEmail" name="email"
+                    <td style="width: 250px;"><input type="email" class="form-control" id="signupEmail" name="email"
                                                      placeholder="이메일주소를 입력하세요."></td>
                 </tr>
                 <tr>
